Fix longestCommonPrefix operating on Unicode
"range" operates on runes, but slicing operates on bytes.

Note that this version does not account for combining characters. For
example, "café" and "cafe" will have a longest common prefix of "caf" if
the 'é' is composed, and "cafe" if the 'é' is decomposed. I'm not sure
it's worth the effort it would take to change this behaviour, especially
since it's unclear what the correct behaviour is.