Check winch before prompt.
If the terminal starts out too narrow, liner never invokes `s.readNext`
meaning that it will never receive the SIGWINCH to indicate that the
terminal has been resized (and doesn't call `s.getColumns` otherwise).
This patch checks the channel before doing the too narrow check and
updates the column size, if there was an update.

Future improvement would be to detect SIGWINCH in `s.tooNarrow` and
abort `s.promptUnsupported`.